TheySell-master/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.1
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
charset-normalizer==2.0.12
cytoolz==0.11.2
Django==3.2.9
django-mathfilters==1.0.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.3.0
eth-typing==2.3.0
eth-utils==1.10.0
frozenlist==1.3.0
gcloud==0.17.0
googleapis-common-protos==1.56.1
hexbytes==0.2.2
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
ipfshttpclient==0.8.0a2
jsonschema==3.2.0
jws==0.1.3
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oauth2client==3.0.0
parsimonious==0.8.1
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
pyparsing==3.0.9
Pyrebase==3.0.27
pyrsistent==0.18.1
python-dotenv==0.19.2
python-jwt==2.0.1
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.7.0
rlp==2.0.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
typing-extensions==3.10.0.2
urllib3==1.26.9
varint==1.0.2
web3==5.25.0
websockets==9.1
yarl==1.7.2
zipp==3.8.0
